move custom git push dialog to dialog package
