+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/actions/RadiclePushAction.java
@@ -10,7 +10,7 @@ import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.util.containers.ContainerUtil;
-import network.radicle.jetbrains.radiclejetbrainsplugin.PushDialog;
+import network.radicle.jetbrains.radiclejetbrainsplugin.dialog.PushDialog;
 import network.radicle.jetbrains.radiclejetbrainsplugin.services.RadicleProjectService;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
